# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  didaliasdto.RegisterTLDDIDParam:
    description: Data Transfer Object for registering a new top-level domain
    properties:
      owner:
        type: string
      tld:
        type: string
    required:
    - owner
    - tld
    type: object
  didaliasdto.RegisterTLDDTO:
    properties:
      did_param:
        $ref: '#/definitions/didaliasdto.RegisterTLDDIDParam'
      priv_key:
        type: string
    required:
    - did_param
    - priv_key
    type: object
...
  response.BlockchainResponse:
    properties:
      errors:
        example: No Error Message
        type: string
      returned: {}
      success:
        example: true
        type: boolean
    type: object
...
  response.Meta:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: DID Backend API
  title: DID API
  version: "1.0"
paths:
...
  /did-alias/did:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the DID hash associated with a node.

        This endpoint returns:
        - The DID hash linked to the node
        - Returns empty hash if no DID is set

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with DID hash
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Node hash (32 bytes hex)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: node
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DID hash
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - malformed node hash
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get DID by Node
      tags:
      - DID Alias GET
  /did-alias/did-from-string:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the DID hash using a domain name.

        This endpoint:
        - Computes the namehash of the domain name
        - Returns the associated DID hash

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with DID hash
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Domain name
        example: example.tld
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DID hash
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - empty name
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get DID by Domain Name
      tags:
      - DID Alias GET
  /did-alias/namehash:
    get:
      consumes:
      - application/json
      description: |-
        Computes the namehash of a domain name according to ENS specification.

        This endpoint:
        - Takes a domain name string
        - Returns its namehash (used as node identifier)

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with namehash
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Domain name
        example: example.tld
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Namehash result
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - empty name
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Compute Namehash
      tags:
      - DID Alias GET
  /did-alias/owner:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the owner address of a node.

        This endpoint returns:
        - The Ethereum address that owns the node/alias
        - Returns zero address if node doesn't exist

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with owner address
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Node hash (32 bytes hex)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: node
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Owner address
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - malformed node hash
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Node Owner
      tags:
      - DID Alias GET
  /did-alias/owner-from-string:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the owner address using a domain name.

        This endpoint:
        - Computes the namehash of the domain name
        - Returns the owner's Ethereum address

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with owner address
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Domain name
        example: example.tld
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Owner address
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - empty name
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Owner by Domain Name
      tags:
      - DID Alias GET
  /did-alias/register-tld:
    post:
      consumes:
      - application/json
      description: Register a new top-level domain in the alias system. Only admins
        can register TLDs.
      parameters:
      - description: Register TLD Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/didaliasdto.RegisterTLDDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered TLD
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    $ref: '#/definitions/response.BlockchainResponse'
                  type: object
              type: object
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new Top-Level Domain (TLD)
      tags:
      - DID Alias POST
  /did-alias/registration-period:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the duration (in seconds) for a given registration period type.

        This endpoint returns:
        - Duration in seconds for the specified period
        - Period types: 0 (ONE_YEAR), 1 (TWO_YEARS), 2 (THREE_YEARS)

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with duration in seconds
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Registration period (0, 1, or 2)
        example: "0"
        in: query
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Registration period duration
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - period must be 0, 1, or 2
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Registration Period Duration
      tags:
      - DID Alias GET
  /did-alias/resolve:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the DID associated with a specific node hash.

        This endpoint returns:
        - The DID hash associated with the given node
        - Returns empty hash if node doesn't exist or has no DID set

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with the DID hash
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Node hash (32 bytes hex)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: node
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DID resolution result
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - malformed node hash or validation error
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Resolve Node to DID
      tags:
      - DID Alias GET
  /did-alias/resolve-from-string:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the DID associated with a domain name string.

        This endpoint:
        - Computes the namehash of the domain name
        - Returns the DID associated with that domain

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): BlockchainResponse with the DID hash
        - `meta` (object): Contains timestamp and API version
      parameters:
      - description: Domain name
        example: example.tld
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DID resolution result
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - empty name or validation error
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Resolve Domain Name to DID
      tags:
      - DID Alias GET
...
swagger: "2.0"